backend/funnel/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# Type inference dependencies
pandas>=2.1.3
numpy>=1.24.3
email-validator>=2.0.0
phonenumbers>=8.13.0

# Google Sheets support
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.116.0
openpyxl>=3.1.2

# Utilities
python-dotenv==1.0.0
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0